fix(user): improve security when updating a user to an email already registered
BREAKING CHANGE: The response from the `PATCH /api/v1/users/[username]` endpoint was modified in
cases where the email is already registered, returning success instead of error. The frontend needs
to handle and inform the user that the confirmation email will only be received if the email
provided is available.